Move settlement utils to own namespace, add documentation comments
