+++ Source/1.3/AI/AIPlayer.cs
@@ -1,9 +1,6 @@
-﻿using RimWorld;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System;
+using Empire_Rewritten.Settlements;
+using RimWorld;
 
 namespace Empire_Rewritten.AI
 {
