+++ Source/1.3/Controllers/FactionController/CivicEthic/EthicAbilityWorker.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Empire_Rewritten.Settlements;
 
 namespace Empire_Rewritten
 {
