+++ Source/1.3/HarmonyPatches/SettlementGizmoPatch.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Verse;
+using Empire_Rewritten.Settlements;
 
 namespace Empire_Rewritten
 {
